+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2.5.1
 * Add InSpec AEM testing support
+* Add support to create AMIs on AWS CodeBuild
 
 ### 2.5.0
 * Fix SSM agent proxy configuration support